updates for metrics
