+++ elastic-otel/src/example-api/StartupExtensions/ServiceDefaultsExtensions.cs
@@ -1,24 +1,17 @@
 ﻿namespace Example.Api.StartupExtensions;
 
-using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using OpenTelemetry;
-using OpenTelemetry.Logs;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Resources;
-using OpenTelemetry.Trace;
 
 internal static class ServiceDefaultsExtensions
 {
-    private readonly static string[] ExcludedOTelRoutes = ["/metrics", "/alive", "/health"];
-
-    internal static IHostApplicationBuilder AddServiceDefaults(this IHostApplicationBuilder builder)
+    internal static IHostApplicationBuilder AddServiceDefaults(this IHostApplicationBuilder builder, string serviceName)
     {
-        builder.AddBasicServiceDefaults("example-api");
+        builder.AddBasicServiceDefaults();
+
+        builder.AddApplicationOpenTelemetry(serviceName);
 
         builder.Services.AddHttpCors(builder.Configuration);
 
@@ -40,7 +33,7 @@ internal static class ServiceDefaultsExtensions
         return builder;
     }
 
-    private static IHostApplicationBuilder AddBasicServiceDefaults(this IHostApplicationBuilder builder, string serviceName)
+    private static IHostApplicationBuilder AddBasicServiceDefaults(this IHostApplicationBuilder builder)
     {
         builder.Services.AddOptions();
         builder.Services.AddHttpClient();
@@ -54,113 +47,6 @@ internal static class ServiceDefaultsExtensions
 
         // Default health checks assume the event bus and self health checks
         builder.AddDefaultHealthChecks();
-
-        builder.AddApplicationOpenTelemetry(serviceName);
-
-        return builder;
-    }
-
-    private static IHostApplicationBuilder AddApplicationOpenTelemetry(this IHostApplicationBuilder builder, string serviceName)
-    {
-        builder.Services.AddLogging();
-
-        Sdk.CreateTracerProviderBuilder()
-            .SetResourceBuilder(ResourceBuilder.CreateDefault().AddService(serviceName))
-            .AddAspNetCoreInstrumentation()
-            .AddHttpClientInstrumentation()
-            .AddOtlpExporter(options =>
-            {
-                options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-                options.ExportProcessorType = ExportProcessorType.Simple;
-                options.Endpoint = new Uri("http://otel-collector:4317");
-            })
-            .Build();
-
-        Sdk.CreateMeterProviderBuilder()
-            .SetResourceBuilder(ResourceBuilder.CreateDefault().AddService(serviceName))
-            .AddAspNetCoreInstrumentation()
-            .AddOtlpExporter(options =>
-            {
-                options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-                options.ExportProcessorType = ExportProcessorType.Simple;
-                options.Endpoint = new Uri("http://otel-collector:4317");
-            })
-            .Build();
-
-        Sdk.CreateMeterProviderBuilder()
-            .SetResourceBuilder(ResourceBuilder.CreateDefault().AddService(serviceName))
-            .AddRuntimeInstrumentation()
-            .AddAspNetCoreInstrumentation()
-            .AddOtlpExporter(options =>
-            {
-                options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-                options.ExportProcessorType = ExportProcessorType.Simple;
-                options.Endpoint = new Uri("http://otel-collector:4317");
-            })
-            .Build();
-
-        var loggerFactory = LoggerFactory.Create(builder =>
-        {
-            builder.AddOpenTelemetry(logging =>
-            {
-                logging
-                    .SetResourceBuilder(ResourceBuilder.CreateDefault().AddService(serviceName))
-                    .AddOtlpExporter(options =>
-                    {
-                        options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-                        options.ExportProcessorType = ExportProcessorType.Simple;
-                        options.Endpoint = new Uri("http://otel-collector:4317");
-                    })
-                    .AddConsoleExporter();
-            });
-        });
-
-        // var openTelemetryBuilder = builder.Services
-        //     .AddOpenTelemetry()
-        //     .ConfigureResource(resource => resource.AddService(serviceName))
-        //     .WithLogging(logging =>
-        //     {
-        //         logging
-        //             .AddConsoleExporter()
-        //             .AddOtlpExporter(options =>
-        //             {
-        //                 options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-        //                 options.ExportProcessorType = ExportProcessorType.Simple;
-        //                 options.Endpoint = new Uri(Environment.GetEnvironmentVariable("OTEL_EXPORTER_OTLP_ENDPOINT") ?? "http://otel-collector:4317");
-        //             });
-        //     })
-        //     .WithTracing(tracing =>
-        //     {
-        //         tracing
-        //             .SetSampler(new AlwaysOnSampler())
-        //             .AddAspNetCoreInstrumentation()
-        //             .AddHttpClientInstrumentation()
-        //             .AddOtlpExporter(options =>
-        //             {
-        //                 options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-        //                 options.ExportProcessorType = ExportProcessorType.Simple;
-        //                 options.Endpoint = new Uri(Environment.GetEnvironmentVariable("OTEL_EXPORTER_OTLP_ENDPOINT") ?? "http://otel-collector:4317");
-        //             });
-        //     })
-        //     .WithMetrics(metrics =>
-        //     {
-        //         metrics
-        //             .AddAspNetCoreInstrumentation()
-        //             .AddHttpClientInstrumentation()
-        //             .AddOtlpExporter(options =>
-        //             {
-        //                 options.Protocol = OpenTelemetry.Exporter.OtlpExportProtocol.Grpc;
-        //                 options.ExportProcessorType = ExportProcessorType.Simple;
-        //                 options.Endpoint = new Uri(Environment.GetEnvironmentVariable("OTEL_EXPORTER_OTLP_ENDPOINT") ?? "http://otel-collector:4317");
-        //             });
-        //     });
-
-        //openTelemetryBuilder.UseOtlpExporter( OpenTelemetry.Exporter.OtlpExportProtocol.Grpc, "")
-
-        // if (!builder.Environment.IsEnvironment("local") && !string.IsNullOrWhiteSpace(Environment.GetEnvironmentVariable("APPLICATIONINSIGHTS_CONNECTION_STRING")))
-        // {
-        //     openTelemetryBuilder.UseAzureMonitor();
-        // }
 
         return builder;
     }
